Disable csslint check for universal selector
Yes, it's known to be slow. But not in all circumstances and the value
it gives, combined with box-sizing, is immense. It's really just slow
when it goes deeper than `* {}`.

However, it also reminded me that the `box-sizing` property really
should be set on the root element (`html`) and inherited from that.

https://github.com/CSSLint/csslint/wiki/Disallow-universal-selector
http://www.paulirish.com/2012/box-sizing-border-box-ftw/